Moved Classroom navbar into a component to clean up Classroom page file
